packages/agents/requirements.txt
google-adk>=1.0.0
google-cloud-texttospeech>=2.16.0
google-generativeai>=0.8.0
playwright>=1.40.0
Pillow>=10.0.0
moviepy>=1.0.3
opencv-python>=4.8.0
numpy>=1.24.0
pydub>=0.25.1
aiohttp>=3.9.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
